requirements.txt
fastapi==0.115.13
PyYAML==6.0.2
uvicorn==0.34.3